app/requirements.txt
aioredis==1.3.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
Deprecated==1.2.13
Django==4.0.4
django-cid==2.2
django-cors-headers==3.11.0
django-oauth-toolkit==1.7.1
django-redis==5.2.0
django-requestlogging==1.0.1
django-splunk-logging==1.1.0
djangorestframework==3.13.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
IMAPClient==2.3.1
incremental==21.3.0
josepy==1.13.0
jwcrypto==1.0
mail-parser==3.15.0
mozilla-django-oidc==2.0.0
msgpack==1.0.3
oauthlib==3.2.0
packaging==21.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.8
python-decouple==3.6
python-json-logger==2.0.2
pytz==2022.1
redis==4.2.2
requests==2.27.1
sentry-sdk==1.5.10
service-identity==21.1.0
simplejson==3.19.1
six==1.16.0
splunk-handler==3.0.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.9
wrapt==1.14.0
zope.interface==5.4.0
